release: integrate 0713

+++ packages/circuits/scripts/copyCircom.sh
@@ -1,4 +1,9 @@
 #!/bin/sh
+# check if circom is installed
+if ! [ -x "$(command -v circom)" ]; then
+  echo 'Error: circom is not installed. Please refer to https://docs.circom.io/getting-started/installation/' >&2
+  exit 1
+fi
 
 rm -rf ./dist/zksnarkBuild
